feat: reading hand compositions!
I'm using a very naive algorithm here, running through a bunch of branches to get a list of possible ways to read a hand. there's probably a lot of space for improvement, but at the same time I'm not sure if that matters: hands are small and generally not overly ambiguous, so in practice I don't expect this to be too heavy.