refactor!: improvements to compose_tiles()
Changes the return type of compose_tiles() to a new struct, which should allow compose_tiles() to be used as part of a planned wait-reading pipeline. I think it also results in more readable code.

This required revising test cases; I took this opportunity to make them a bit more concise.